% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfs_functions.R
\name{clean_cfs_variables}
\alias{clean_cfs_variables}
\title{Clean Clinical Frailty Scale Variables}
\usage{
clean_cfs_variables(df, variables, allowed_values)
}
\arguments{
\item{df}{A data frame.}

\item{variables}{A character vector of column names to be cleaned.}

\item{allowed_values}{A vector of valid values for the variables.}
}
\value{
The data frame with the specified variables cleaned.
}
\description{
Cleans specified variables in a data frame, replacing invalid values and '9' with NA.
}
